// src/lib/types.ts
export type VideoMediaAction = 'direct-download' | 'merge-then-download' | 'browser-hls-download' | 'hide'
export type AudioMediaAction = 'direct-download' | 'extract-audio' | 'hide'

export interface MediaActions {
    video: VideoMediaAction;
    audio: AudioMediaAction;
}
// ...
export interface UnifiedParseResultImage {
    index?: number;
    url?: string | null;
    downloadUrl?: string | null;
}